+++ src/features/home/index.ts
@@ -1,1 +1,3 @@
+export * from "./hooks/useHomeDashboard";
 export * from "./pages/HomePage";
+export * from "./types";